2/example/requirements.txt
attrs==19.3.0
base58==2.0.0
bitarray==1.2.1
certifi==2020.4.5.1
chardet==3.0.4
cytoolz==0.10.1
eth-abi==2.1.1
eth-account==0.5.2
eth-hash==0.2.0
eth-keyfile==0.5.1
eth-keys==0.3.3
eth-rlp==0.1.2
eth-typing==2.2.1
eth-utils==1.9.0
hexbytes==0.2.0
idna==2.9
importlib-metadata==1.6.0
ipfshttpclient==0.4.13.2
jsonschema==3.2.0
lru-dict==1.1.6
multiaddr==0.0.9
netaddr==0.7.19
parsimonious==0.8.1
protobuf==3.11.3
pycryptodome==3.9.7
PyJWT==1.7.1
pyrsistent==0.16.0
requests==2.23.0
rlp==1.2.0
six==1.14.0
toolz==0.10.0
typing-extensions==3.7.4.2
urllib3==1.25.9
varint==1.0.2
web3==5.9.0
websockets==8.1
zipp==3.1.0
